requirements.txt
streamlit
geopandas
shapely
rasterio
sentence-transformers
faiss-cpu
transformers
torch
pandas
Pillow
requests
